Move PoS reward functions to gridcoin/staking directory
